% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_example.R
\name{fp_example}
\alias{fp_example}
\title{Get path to fpod example}
\usage{
fp_example(path = NULL)
}
\arguments{
\item{path}{Name of the file. If NULL, all example files are listed.}
}
\value{
A character vector, containing the path(s) that matched, or the
empty string, "", if none matched.
}
\description{
Get path to fpod example
}
\examples{
# get a list all example files
fp_example()

# get the path for this particular file
fp_example("gullars_period1.FP3")

}
