% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_find_buzzes.R
\name{fp_find_buzzes}
\alias{fp_find_buzzes}
\title{Finds harbor porpoise feeding buzzes}
\usage{
fp_find_buzzes(x, method = "clicks")
}
\arguments{
\item{x}{a data.table where each row is a click, as the "clicks" element in
the list object returned by \code{\link[=fp_read]{fp_read()}}. Each row must minimally have a
POSIXct column \code{time}, with nanosecond precision.}

\item{method}{the method to use to find feeding buzzes - "clicks" or "trains". See details.}
}
\value{
An integer vector of the same length as \code{nrow(x)}, where the values
indicate that that click can be considered a feeding buzz (value = 1) or not (value = 0).
}
\description{
This function uses one of two methods to classify NBHF clicks into one of
two classes: feeding buzz or non feeding buzz.
}
\details{
Note that the so-called "feeding buzzes" are usually considered to represent
a combination of feeding buzzes and social calls. Even so, the classifications
resulting from both of these methods are commonly used as a proxy of foraging
activity.

The two available methods are:
\itemize{
\item \code{clicks} method: any inter-click-interval (ICI) less than 20 milliseconds
is considered a NBHF feeding buzz.
\item \code{trains} method: buzzes are identified using a mixture Gaussian model, with
the number of components k=3. All clicks associated with the first component
are considered a NBHF feeding buzz. This method requires the package mixtools
to work.
}
}
\examples{
# first read some FPOD data
fn <- fp_example("gullars_period1.FP3")
dat <- fp_read(fn)

# extract porpoise clicks of quality Hi and Mod
nbhf <- dat$clicks[species == "NBHF" & quality_level >= 2]

# then add a 'feeding buzz' column to the clicks data.table
nbhf$buzz <- fp_find_buzzes(nbhf, method = "clicks")

}
\references{
Pirotta, E., Thompson, P.M., Miller, P.I., Brookes, K.L., Cheney,
B., Barton, T.R., Graham, I.M. and Lusseau, D. (2014), Scale-dependent foraging
ecology of a marine top predator modelled using passive acoustic data. Funct
Ecol, 28: 206-217. https://doi.org/10.1111/1365-2435.12146
}
