% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{fp_floor_timestamp}
\alias{fp_floor_timestamp}
\title{Internal helper function to floor timestamps to the nearest desired time unit}
\usage{
fp_floor_timestamp(x, unit)
}
\arguments{
\item{x}{a vector of POSIXct timestamps}

\item{unit}{character denoting the precision desired, one of \code{minute}, \code{hour}
or \code{day}}
}
\value{
a vector of rounded timestamps
}
\description{
Internal helper function to floor timestamps to the nearest desired time unit
}
\keyword{internal}
