% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_summarize.R
\name{fp_summarize}
\alias{fp_summarize}
\title{Calculates minute-resolution summaries of detections and other variables}
\usage{
fp_summarize(clicks, env, header)
}
\arguments{
\item{clicks}{data.table where each row is a click, as the "clicks" element in
the list object returned by \code{\link[=fp_read]{fp_read()}}. Each row must minimally have a
POSIXct column \code{time}.}

\item{env}{data.table, as returned from \code{\link[=fp_read]{fp_read()}}}

\item{header}{list, as returned from \code{\link[=fp_read]{fp_read()}}}
}
\value{
A data.table with three or four columns:
\itemize{
\item timestamp: POSIXct timestamp of the start of the 1-minute time chunk, in YYYY-mm-dd HH:MM format
\item degC: temperature in degrees Celcius, as reported by the POD.
\item dpm: detection-positive-minutes, 1 if at least one click is registered during the time chunk; 0 otherwise.
\item bpm: buzz-positive-minutes, 1 if at least one feeding buzz is registered during the time chunk, 0 otherwise.
Note that bpm is only available if there is a 'buzz' column in the clicks data.table.
}
}
\description{
Calculates minute-resolution summaries of detections and other variables
}
\details{
The reason that \code{clicks}, \code{env} and \code{header} must be specified separately
(rather than just directly passing the return value from \code{\link[=fp_read]{fp_read()}}) is to
allow subsetting only the clicks of interest, e.g. NBHF of quality 2 and 3.
}
\examples{
# first read some FPOD data
fn <- fp_example("gullars_period1.FP3")
dat <- fp_read(fn)

# extract porpoise clicks of quality Hi and Mod
nbhf <- dat$clicks[species == "NBHF" & quality_level >= 2]

# get a simple summary, with timestamp, degC and dpm.
dpm <- fp_summarize(nbhf, env = dat$env, header = dat$header)

# If we also wanted buzz-positive-minutes, we could do this first.
nbhf$buzz <- fp_find_buzzes(nbhf, method = "clicks")
# This time, when we call fp_summarize, it should detect that there's a
# buzz column, and automatically calculate feeding-positive-minutes as well.
dpm <- fp_summarize(nbhf, env = dat$env, header = dat$header)

# Once we have DPMs, we can easily aggregate these into coarser time chunks
dpm_per_hour <- dpm[, .(dpm = sum(dpm)),
    .(date = fpod:::fp_floor_timestamp(timestamp, unit = "hour"))] # per hour
dpm_per_day <- dpm[, .(dpm = sum(dpm)),
    .(date = as.Date(timestamp))] # per day
dpm_per_week <- dpm[, .(dpm = sum(dpm)),
    .(date = fpod:::fp_floor_timestamp(timestamp, unit = "week"))] # per week

}
\seealso{
\code{\link[=fp_find_buzzes]{fp_find_buzzes()}}, \code{\link[lubridate:round_date]{lubridate::floor_date()}}
}
