% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fpod.R
\name{read_fpod}
\alias{read_fpod}
\title{Read FPOD data}
\usage{
read_fpod(file, tz = "", simplify = TRUE, amp = "extended")
}
\arguments{
\item{file}{a character string. The path to the FPOD (or CPOD) data file.}

\item{tz}{a character string. The time zone specification to be used for
calculating dates. Passed unchanged to \code{\link[=as.POSIXct]{as.POSIXct()}}.}

\item{simplify}{logical. If TRUE, simplifies the clicks data.table by stripping
away some columns, such as \code{clk_ipi_range}, \code{ipi_pre_max}, \code{amp_reversals},
\code{duration}, and \code{has_wav}.}

\item{amp}{a character string. With \code{amp}="extended", higher values are extrapolated
from the duration of clipping and the IPI. For any other values of amp, the
compressed SPL values recorded by the FPOD are used directly.}
}
\value{
A list, with one or more of the following data.frames (or data.tables, if available):
\itemize{
\item header: a list with pod name, coordinates, starting time, stopping time, user notes, etc.
\item clicks: A data.frame (or data.table) with data about each click. See details.
\item wav (only FPx files): pseudo-wav data - inter-peak-intervals and raw amplitudes for a subset of clicks
\item env: misc data, ambient temperature (in deg C) and battery voltage per stack (in units of 10 millivolts)
}
}
\description{
This function reads an FPOD or CPOD data file (FP1, FP3, CP1, CP3) into R.
}
\details{
The clicks data.frame contains the following columns:
\itemize{
\item time: The time and date of the click, at microsecond resolution. Note that R might
only display dates and times to a second precision, but any date or time calculations
will use the full precision.
\item minute: minutes elapsed, since starting the FPOD
\item microsec: microseconds elapsed, since the start of the minute.
\item click_no: an ID number that uniquely identifies the click.
\item train_id: an ID number from the KERNO classifier, reset for each minute.
\item species: the species classification from the KERNO classifier
\item quality_level: the quality level of the classification, 1 to 3.
\item echo: TRUE if the KERNO classifier thinks this click might be an echo of another click that has already been classified.
\item ncyc: number of cycles in click. This is a proxy for the click duration.
\item pkat: the number of the cycle with the highest amplitude
\item clk_ipi_range: range of inter-peak-intervals (IPIs) among cycles in click
\item ipi_pre_max: the IPI before pkat, in 250 nanosecond units
\item ipi_at_max: the IPI at pkat, in 250 nanosecond units
\item khz: The frequency (in kHz) of the peak cycle.
\item amp_at_max: the peak amplitude (SPL) of the loudest cycle
\item amp_reversals: the number of amplitude reversals
\item duration:
\item has_wav: TRUE if there is a pseudo-WAV recorded for this click.
}
}
\examples{
# read a FP3 file
dat <- read_fpod(file = "helga period 1.FP3")

# show misc. information (pod number, deployment date, etc.)
dat$header

# tally up the number of clicks in each species category
table(dat$clicks$species)

# Calculate Detection Positive Minutes (DPMs) for porpoises per day
 dpm_per_day <- dat$clicks[species=="NBHF", length(unique(minute(time))), as.Date(time)]
}
