% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_summarize.R
\name{fp_summarize}
\alias{fp_summarize}
\title{Calculates minute-resolution summaries of clicks}
\usage{
fp_summarize(x)
}
\arguments{
\item{x}{data.table where each row is a click, as the "clicks" element in
the list object returned by \code{\link[=fp_read]{fp_read()}}. Each row must minimally have a
POSIXct column \code{time}. The return value of \code{\link[=fp_read]{fp_read()}} is also accepted with a warning,
provided that the clicks data.table is present .}
}
\value{
A data.table with three or four columns:
\itemize{
\item time: POSIXct timestamp of the start of the 1-minute time chunk, in YYYY-mm-dd HH:MM format
\item dpm: detection-positive-minutes, 1 if at least one click is registered during the time chunk; 0 otherwise.
\item bpm: buzz-positive-minutes, 1 if at least one feeding buzz is registered during the time chunk, 0 otherwise.
Note that bpm is only available if there is a 'buzz' column in the clicks data.table, e.g.
from first calling \code{\link[=fp_find_buzzes]{fp_find_buzzes()}}
}
}
\description{
Calculates minute-resolution summaries of clicks
}
\examples{
# first read some FPOD data
fn <- fp_example("gullars_period1.FP3")
dat <- fp_read(fn)

# extract porpoise clicks of quality Hi and Mod
nbhf <- dat$clicks[species == "NBHF" & quality_level >= 2]

# get a simple summary, with timestamp, degC and dpm.
dpm <- fp_summarize(nbhf)

# If we also wanted buzz-positive-minutes, we could do this first.
nbhf$buzz <- fp_find_buzzes(nbhf, method = "clicks")

# This time, when we call fp_summarize, it should detect that there's a
# buzz column, and automatically calculate feeding-positive-minutes as well.
dpm <- fp_summarize(nbhf)

# Once we have DPMs, we can easily aggregate these into coarser time chunks
dpm_per_hour <- dpm[, .(dpm = sum(dpm)),
    .(date = as.POSIXct(trunc(time, unit = "hours")))] # per hour
dpm_per_day <- dpm[, .(dpm = sum(dpm)),
    .(date = as.Date(time))] # per day
dpm_per_month <- dpm[, .(dpm = sum(dpm)),
    .(date = as.POSIXct(trunc(time, unit = "months")))] # per month

}
\seealso{
\code{\link[=fp_find_buzzes]{fp_find_buzzes()}}, \code{\link[lubridate:round_date]{lubridate::floor_date()}}
}
