% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpod-package.R
\docType{package}
\name{fpod-package}
\alias{fpod}
\alias{fpod-package}
\title{fpod: Read and process FPOD and CPOD data}
\description{
This package reads FPOD and CPOD data into R directly from the FPOD data files (i.e. the .CP1, .CP3, .FP1 and .FP3 files). The FPOD data files contain binary data, so they can't trivially be read into R using the usual approach (e.g. fread or read.csv). This package decodes the binary data and imports all the data in one go (i.e. header/metadata, clicks, KERNO classifications, environmental data and pseudo-WAV data). It is then trivial to aggregate data as you please, e.g. DPMs per time block. The advantage of handling data processing in R is a long topic, but suffice it to say that it 1) simplifies things (many fewer steps, as different vars have to be exported in multiple goes in the official FPOD app), and more importantly, 2) makes data processing 100% transparent and reproducible.
}
\seealso{
\code{\link[=fp_read]{fp_read()}}, \code{\link[=fp_summarize]{fp_summarize()}}, \code{\link[=fp_find_buzzes]{fp_find_buzzes()}}
}
\author{
\strong{Maintainer}: André Moan \email{andre.moan@hi.no}

}
\keyword{internal}
